+++ src/endstone/core/command/console_command_sender.h
@@ -26,6 +26,8 @@ public:
     void sendErrorMessage(const Message &message) const override;
     [[nodiscard]] std::string getName() const override;
     [[nodiscard]] PermissionLevel getPermissionLevel() const override;
+    [[nodiscard]] bool hasPermission(std::string name) const override;
+    [[nodiscard]] bool hasPermission(const NotNull<Permission> &perm) const override;
 };
 
 }  // namespace endstone::core